Update patch set 6
Patch Set 6:

You should consider only bash commands skipping all lines that doesn't start with a plus. This i something easy to parse with your script (it has been cleanup):

http://osjenkins.ir.intel.com:8080/view/fressi/job/networking-odl-ci/ws/build/90/console.txt

The one with ++ are sub operations of the ones with one +, therefore the time of the ones with one + should extend until the next one with only one +. Y could modify my script to do also that and two save statistics for calls on a file.

This is the script I think to improve with this feature:
https://github.com/FedericoRessi/devstack-ws/blob/master/scripts/bash.py

You can use it to run devstack. It is just a wrapper of the bash. It stores logs files on a directory set by environment variable LOG_DIR. It executes a command capturing stdin and stdout. It saves this streams do 3 log files: .ansi is coloured text, .txt has ansi colours removed, .html was processed by ansi2html.
All logs i preprocessed to remove initial date and it writes the date when de log is line is received. Therefore for computing time differences should be OK.

Patch-set: 6
Reviewer: Gerrit User 17377 <17377@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0