Update patch set 3
Patch Set 3: Code-Review-1

Do not remove the note about creating an SSH key and adding it to all nodes so the user can ssh in. That is still required information. 

Also, can we place all this information about configuring SSH keys for live migration in its own section? The old section mostly was just about creating an SSH key so the user can ssh in to each node.

There are sections for misc:

http://docs.openstack.org/developer/devstack/guides/multinode-lab.html#options-pimp-your-stack

or

http://docs.openstack.org/developer/devstack/guides/multinode-lab.html#notes-stuff-you-might-need-to-know

Patch-set: 3
Label: Code-Review=-1