Cannot use pip_install_gr before install_infra
The library os-testr cannot be installed using pip_install_gr (line
693). In fact the global-requirements.txt file is downloaded by the lib
infra, which is installed later in the file (line 715).

As done in master branch, it is possible to use `pip_install -U` instead
of `pip_install_gr`. I do not suggest to install infra before for
consistency with the other branches.

Change-Id: I632cb83fab264cde8edb5319b93809e11e4f2352
Closes-Bug: #1539748