Update patch set 2
Patch Set 2:

> Enrico, I am not sure I understood what you said, but we do use
 > pip_install_gr which do use the requirements that were just cloned.

Now it is correct (it was modified in the second patch set).

By the way the command:
git_clone $REQUIREMENTS_REPO $REQUIREMENTS_DIR $REQUIREMENTS_BRANCH
seems to be one of the tasks of the install_infra module, so why should we duplicate it also here?

Second problem for me: in master and stable/liberty the line is pip_install -U os-testr. Should i submit a patch to change those? stable/kilo is a prior version, it should not contain better code wrt newer branches.

Patch-set: 2