Use pip_install_gr for os-testr
Use `pip_install_gr` instead of `pip_install -U` when installing
os-testr. This is the same approach that have just been approved in
the stable/kilo branch (https://review.openstack.org/#/c/274533/).

This is an improved way to install depencencies that do not conflict
with other ones defined in the global requirements repository.

Change-Id: I999faaabfdcf4530a1f0acb4906160fa027442a6