Use pip_install_gr for os-testr
Use `pip_install_gr` instead of `pip_install -U` when installing
os-testr. This is the same approach that have just been approved in
the stable/kilo branch (https://review.openstack.org/#/c/274533/).

This is an improved way to install depencencies that do not conflict
with other ones defined in the global requirements repository.

Change-Id: Iab4298b148ca68091e5201a2c24ecbbf4cefa63f