Update patch set 1
Patch Set 1: Code-Review-1

I think this it is obvious that you must clone the devstack repo with a user that has appropriate permissions for the parent directory, as well as the devstack directory.

The only time this doesn't happen, is if someone is logged in as root and tries to do everything as root. Which is bad practice. I don't think we need to state this in the docs.

Patch-set: 1
Reviewer: Gerrit User 4656 <4656@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1