Enable showing multiple locations for ceph
Currently glance v2 doesn't allow to specify custom
locations for image by default, it returns 403.
To enable this 'show_multiple_locations' param must
be set to True.

When Nova doing snapshots with Ceph it provides location
to activate image, which fails when we use Glance v2.

To fix it we change the default param to True.
'show_image_direct_url' will be overridden.

Related to blueprint use-glance-v2-api

Change-Id: I2bd454990d90f60b0d6505d56c9ce23836e82e3b