Update patch set 1
Patch Set 1:

If this is blocking in the gate because of a patch then we should have it here and in the plugin (we are currently in the process of changing the voting job to use the plugin). If not, then abandon this change and submit it in the plugin devstack :).

Thanks!

Patch-set: 1