Update patch set 3
Patch Set 3:

Not really opposed, just like to be sure we're making sure we are doing it in the best way and not just putting noise in the worlddump that isn't that useful and will never go away.

I just feel like with devstack, these will all be shut down before we call this.

There's other calls like [1] where this is noise because it's running *before* neturon, I guess (check the logs [2,3])

for example, maybe it would just be better to do this in neutron shutdown scripts all the time?

[1] https://git.openstack.org/cgit/openstack-dev/grenade/tree/projects/60_nova/resources.sh#n98
[2] http://logs.openstack.org/35/279035/3/check/gate-grenade-dsvm-neutron/b3b7e17/logs/worlddump-2016-02-12-114736-nova_resources_created.txt.gz
[3] http://logs.openstack.org/35/279035/3/check/gate-grenade-dsvm-neutron/b3b7e17/logs/worlddump-2016-02-12-114822-cinder_resources_created.txt.gz

Patch-set: 3