Fix devstack from dying on unsupported Distros
I believe it is not the correct behavior to die on unsupported distros.
Removing the ability to stack on certain distros even when checks resides for
distros such as LinuxMint is not the way to go about this IMO. We should support
LinuxMint or allow the catchall and let the deployers deal with it. If complaints
arise regarding 'Why doesnt it work for my distro' they should be
pointed to documentation regarding the supported distro list.

Change-Id: I42599c1bd851d2b61a56a3960f42ad051aab1727
Closes-Bug: #1545864