Update patch set 1
Patch Set 1:

Sorry, was saying I think I see what Ian / previous commit was going for --limiting liability and stop people from running on totally unsupported distros, then filing a ton of invalid bugs and wasting support time. That makes a ton of sense. It's just lame that there's no easy way to say "YEAH, I KNOW, BUT THIS WORKS" and move on, and thus causes a nearly equal amount of frustration (except this time, it's for the developer who is actually on a working system being arbitrarily blocked).

Patch-set: 1