Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

terribly minor nit but it's going to fail bashate

this turned out to be quite a bit more complicated than i'd hoped :(  do we have to run this on a proposal slave -- we could move it to a full node?  doesn't seem to matter too much but we do prefer to not rely on network so much for ci jobs.

maybe for simplicity we should just remove the other path and just leave the json query?  is it really going to do anything other than bit-rot?

Patch-set: 1
Label: Code-Review=-1