Update patch set 2
Patch Set 2:

You and me both.  I kept the other path because a) it's way faster locally and gets the date as with Matt's original code b) there's a comment in proposal_update.sh that suggests a persistent git cache will come to the proposal slave someday.  OTOH it could easily be dropped now and resurrected if that ever happens.

As for running on a full node, I don't know if that's an option.

Patch-set: 2