Make 'public' network the default external network
Neutron implemented an extension to allow users to automatically
provision a basic network topology to connect their instances.

One of the requirements for this feature is to be able to mark
an external network to be the one to be used for external
connectivity.

This patch adds the flag needed to codify this requirement.

Change-Id: I43ce5d65e754f131f7ca1ce2088a397d266cf821
Partially-implements: blueprint get-me-a-network