Demote check_libs_from_git to non-fatal
check_libs_from_git verifies that the names listed in LIBS_FROM_GIT are
in fact pip-installed things that look like they were installed from
git.  Previously, an unrecognised name resulted in a fatal error - this
change demotes that to a warning (and the unrecognised name is
effectively ignored).

This makes fat-fingered values less obvious, but opens the possibility
of auto-generating LIBS_FROM_GIT without knowing a-priori whether a git
repo name is going to be a pip-installed thing.

Change-Id: Ic35b4cbfef6e19d2322bd19e0e31312c48642ecd