Update patch set 2
Patch Set 2: Code-Review+1

Looks good. The only change of actual exported variables is in openrc and the old value is maintained. 

The only possible issue i see is if someone is passing in their own SERVICE_TENANT_NAME variable it's lost - but i see this as so unlikely we should merge it and fix it if someone complains.

Patch-set: 2
Reviewer: Gerrit User 7191 <7191@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1