Disable tempest tox venv recreation in OFFLINE mode
Setting OFFLINE=True in local.conf should enable stack.sh to run multiple
times without an internet connection. This was broken for some cases
when recreation of tox venv for tempest was forced in lib/tempest.

This change makes recreation of tox venv enabled only when OFFLINE mode is
disabled.

Change-Id: I2bf6caf60038a3690378eead905b35e9175ac356
Closes-Bug: #1550537