Work-around pre-installed pip packages on Fedora
Work-around some packages being pre-installed with pip as part of the
Openstack CI image build for Fedora.  We need to clear these out,
otherwise we try to install a packaged version over the top, which
fails, and then we try to run fixups in fixup_stuff.sh which fails,
and basically it is a big mess.

Change-Id: I6b06dcd897999e0642b387e8d13a473934ed0956