Update patch set 3
Abandoned

This hasn't had any activity in almost a year and failed jenkins. (also ceph has been moved to a devstack plugin) So, I'm abandoning this patch. If you'd like to continue working on it, feel free to open this change against the ceph plugin and bring it up to date.

Patch-set: 3
Status: abandoned