Update patch set 1
Patch Set 1: Code-Review-1

stop_nova_compute got it's pid from the pid file

 2016-02-09 20:12:13.302 | ++ [[ -r /opt/stack/status/stack/n-cpu.pid ]]
 2016-02-09 20:12:13.302 | +++ cat /opt/stack/status/stack/n-cpu.pid
 2016-02-09 20:12:13.304 | ++ pkill -g 12999
 2016-02-09 20:12:13.311 | + exit_trap
 2016-02-09 20:12:13.311 | + local r=1
 2016-02-09 20:12:13.311 | ++ jobs -p
 
except nova had bailed out before that, i guess leaving a stale pid file [1]

at the root of this, stop_process is not safe when running under "set -e" (errexit).  so this is not really going to do anything; it seems the point is to restart a broken nova that has failed and left a stale pid-file, but stop_process will just halt things when the pkill fails (as noted in prior comment)

i'd have to think about whether this is intentional or not.  if we're trying to stop something that has a stale pid file, maybe we really want to bail out, as clearly something is not as we intended.  my first thought it maybe we pass down a flag that says "stop this, but i don't care if it is already failed".

so, before this restart can do what is intended, I think we need to consider the semantics of stop_process and fix that so it can work in this case.

[1] http://logs.openstack.org/91/271391/2/check/gate-tempest-dsvm-ironic-pxe_ipa-ipxe/2fe3646/logs/screen-n-cpu.txt.gz#_2016-02-09_20_11_21_954

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1