Fix issues with test_package_ordering unit test
The test_package_ordering.sh test is failing intermittently because it
doesn't always ignore case, or use a dictionary sort.

This patch changes the sort to use dictionary sort and ignore case (-df)
which should ensure the test is more consistent, regardless of sort
version, and other inconsistencies.

Additionally the rpms-suse/horizon file order was changed to abide by
this consistency.

Change-Id: I5ba46a7faa837208a98c26b4831673e7c7f43c88
Closes-Bug: #1554488