[WIP] Neutron refactor: Move L3 crud into separate service file
There's a lot of L3 configuration that was in the main neutron file, but
a lot of it is self contained and can be moved into its own file.

The new l3 service file will contain all the previous L3 plumbing and
configuration that the OpenStack Gate expects, while also eventually
moving the whole l3 network creation step into a single hook that can be
overridden by plugins.

Change-Id: I8d92089a9aa02c72af72806a5a9427a77b30673e