Allow ironic scenario's power_timeout to be set by localrc option
Currently this value is not set by devstack and the default from
tempest is used. While the default (currently 10 seconds) is
viable for testing libvirt guests as targets, real hardware can
take longer to power on. This patch adds the option to set the
power_timeout for the baremetal scenario via a localrc option.

Change-Id: I2c5f22c6c7274e126453f609a550ce38fecc3e75
Closes-bug:#1555176