Update patch set 1
Patch Set 1:

(3 comments)

Firstly, this does not seem bad.

A lot of it is about installing virtualbox and setting up and ubuntu image.  I wonder if there are better guides to that which will not go out-of-date.

My feeling on the direction for this : I feel like it might be better to start from one of the ubuntu cloud images.  I know they are qcow2, but you can fairly easily convert that to virtualbox if you want to keep that direction (I'm pretty sure qemu-img even got support to convert to vdi?).  

At that point, you can describe how to create a small iso to inject the meta-data for ssh and the user-script for the "shake-and-bake" bit we have in the existing documentation (I wrote a post on this ages ago [1]).  That would, I imagine, work just as well for parallels, fusion, whatever else people are using, so be a bit more generic.  

This way, we're getting new users up to the point they're working in a image booting cloud-init, and from that point instructions are common no matter what's happening.  Getting to the point the "shake-and-bake" instruction takes over seems like something concrete we can maintain. 

What do you think?

[1] https://www.technovelty.org/linux/running-cloud-images-locally.html

Patch-set: 1
Label: Verified=0