Update patch set 2
Patch Set 2:

various package managers (yum, apt-get) have different parameters and interfaces and they don't exactely overlaps. Therefore we should focus on having functions to abstract them (as now) instead of using variables as you proposed.
By the way for the specific case of Fedora / Red Hat it is fine to use a variable as you proposed because DNF is expected to be backward compatible with YUM.
You could hava a variable called YUM that in the case of Fedora 22+ it is set to 'dnf' instead of 'yum'. As I said before I would avoid looking at the specific version (22, 23 ,etc.) because are informations binded to a specific vendor release. You should instead look at if dnf is installed and working. You would have a code easier to reuse and independently from Red Hat release strategies.
Please also consider that dnf could be backported to older distributions and it should add reasonable speed improvements over yum.

Patch-set: 2