Update patch set 1
Patch Set 1:

Thanks Ian for your feedback. I agree there may be better ways, however this resulted from the first hand experience of the bugsmash.  It takes the lowest components of use, and provides a step by step guide. It's not as automated as I would generally use, but if the person installing this has not used virtualization or Linux a lot (or even at all), it serves as a grounding on some of those basics we never consider after we know what we are doing. (e.g. doing the hard yards about a network interface) I don't even talk much about it, but it's needed to use devstack effectively when you know more.

This review is a result of "rather than blog about, add it to the docs so there are not so many google hits on out of date devstack installation guides" comment. 

I did think about breaking it up, having a later section, but honestly this is how you do it the first time with no experience, why not be specific and have it in one place.  I suspect this will not go out of date much at all, it is basically how I installed this the first time myself over a year ago.  Also being in one place its easier then a blog for anybody to keep it up to date.

I'll update my docs with your other recommendations.

Patch-set: 1
Label: Verified=0