Update patch set 1
Patch Set 1:

(1 comment)

Sorry, I forgot that you can disable service with '-'. My -1 still remains and mostly due to missing ','. 

It is still nice to be able to clear all defaults services without having to add '-' to every one of them (current behavior). But that is just my opinion.

Patch-set: 1