Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

I am with Ian here and would prefer to not see this level of guide in the DevStack repo.  Specifying a 'running Ubuntu server VM' should be enough.  There is a certain minimum competency required to operate/develop OpenStack and I think creating a server VM is included in that.

If we do add this, however, I really do not want to see others added, one additional bit-rotting guide is enough. We can be opinionated here and this will be that opinion on local DevStack VMs.

I am also on board with starting with a Trusty cloud image, it will save a bunch of time, especially if you start with 'current'.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1