[WIP] Rabbit - Lower retry timeout for missing destination
There was an attempt before:
https://review.openstack.org/#/q/8c03a6db6c0396099e7425834998da5478a1df7c,n,z

But the lower threshold was just too low (0.5). We should set it to
1/2 the time of the the rpc_response_timeout which is set to 60 seconds.

In the specific bug, nova api calls nova compute which calls nova
network which fails because a consumer went AWOL (when grenade stopped
the old side and started the new side). Since nova network waited
the full 60 seconds the rest of the previous calls time out. At least
with 30 seconds, there's a chance for the retry to succeed and
the responses flow back.

Closes-Bug: #1545002
Change-Id: I94a1e7393a3ccfc7c2cdd7f1489644b50a582d4b