Update patch set 2
Patch Set 2:

@Dean, You cannot  assume "certain minimum competency required to operate/develop OpenStack and I think creating a server VM is included in that." 

The production of this document was a direct result of attending an OpenStack BugSmash day, not even an introduction event.  After helping five people create a suitable VM, it said I'll create a blog post for doing this. The response was we don't need to add to the numerous inaccurate or out of date blog posts, we really want to direct people to devstack documentation. To which I would agree.

If we want to set a minimum skill level then we should direct people to applicable documentation to achieve said minimal skill if they do not have it.  

I was also unaware of the simplicity of using a cloud image which I will try out, however this again highlights there are alternative approaches people use.

Patch-set: 2
Label: Verified=0