Update patch set 1
Patch Set 1: Code-Review-1

I think this is a bit too much "magic" configuration.

To take a step back - where is all of this work going? I'm counting a huge amount of OVS knobs that we are now creating, and frankly I am starting to get less interested in this, as more patches come in. I will comment on the other patch to the same effect.

Patch-set: 1