Update patch set 1
Patch Set 1:

one thing that i have been thinking about for the last 6 months is that there is a need
to have a single libary in devstack that can install and configure ovs which can be used
when deploying ovs with the neuton agent, odl or ovn. 
currently both networking-ovn and networking-ovs-dpdk have different code to compile and install ovs from source. 

networking-ovs-dpdk can be used to deploy ovs with dpdk with the neutron agent.
the networking-ovs-dpdk plugin can be combined with the networking-odl plugin also to have odl mangage ovs. 
networking-ovn is currently not compatible with networking-ovs-dpdk but we hope to fix that in the N cycle.
it would be a lot simpler if we could add a lib to devstack to handle ovs installation(source and package) and bridge creation/port addtion.
would that be a perferable option to these minor changes? 

it would appear that neutron has recently added support for compiling ovs from source to test connection tracking.
https://github.com/openstack/neutron/blob/master/devstack/lib/ovs

perhaps we can contribute our changes as part of that effort instead if you do not want to have them in devstack directly.

Patch-set: 1