Update patch set 2
Patch Set 2:

i do think it would be good too have a single ovs plugin that can deploy ovs from source or binary with and without dpdk that handeled just the installation of ovs and the configuration of its bridges and ports.

if that plugin was created it would remove the need for the networking-ovs-dpdk devstack pluging which does most of this and could be reused with networking-ovn and networking-odl.
it would also allow all ovs specific logic to be removed form
devstack. similarly the same could be done for linux bridge if
desired.

if such a plugin was created would it make more sense to create a networking-ovs repo for that or to extend the neutron devstack plugin?
if the reference plugins had been split out as originally intended then i think that having a networking-ovs repo which contained the ovs agent, mechanism driver and devstack plugin would make sense but as that is still in tree would it make sense to move this logic to the neutron devstack plugin https://github.com/openstack/neutron/tree/master/devstack

Patch-set: 2