Update patch set 1
Patch Set 1:

@Sean Mooney

I don't believe the solution is to put it in the Neutron devstack plugin. That plugin has a lot of stuff in it, and it has seen organic growth. I would *prefer* that if there really are a lot of OVS specific things (such as compiling OVS from source, doing bridge plugging, and so forth), that it be an *OVS* specific devstack plugin, rather than being placed into other existing devstack plugins that just bloat things up more.

Patch-set: 1