Add can_branch_stable.py tool for detecting release edges
This adds a tool that lets you check to see if all dependent projects
have cut stable branches yet. This should help in determining when we
can branch stable devstack.

In order to do this, I pulled the list of projects out into lib/projects
and use that both to source the project libraries and to use for checking
github in the tool.

Change-Id: I9c71bb39283ef178a8eb8bf2b805ca0cd051237d