Update patch set 1
Patch Set 1:

but, there's also a post-install phase

so, it's not totally clear where this should be.  I guess technically it's not required for neutron so probably it should be in that post-install phase.  this needs more discussion

[1] https://git.openstack.org/cgit/openstack/neutron/tree/tools/configure_for_func_testing.sh#n217

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0