Ensure testing configuration can run as late as possible
Before the code in the extra.d plugins was removed from the devstack
tree they could define the order they ran. When this code is decomposed
into a plugin, there is still a need to do some form of ordering. This
caused problems with the Ironic devstack plugin and Tempest because the
code is run in this order:

 1. The tempest configuration is run from extra.d, processing
    DEFAULT_INSTANCE_TYPE, and writing the flavor_ref to tempest.conf

 2. The Ironic devstack plugin is run, creating the flavor needed for
    DEFAULT_INSTANCE_TYPE

This leads to build failures as tempest can not find the required
flavor, so it writes which ever flavor it can find at the time into
flavor_ref. Ironic now has code it its devstack plugin duplicated from
the tempest plugin to work around this problem until this is merged.

This patch fixes this by making the post-extra phase useable by the
extra.d and out of tree plugins. This means that tempest and any other
plugin that needs really late configuration can use this stage to wait
until all other processes are done configuring to do their
configuration. However it still runs before the local.conf configuration
overrides.

Change-Id: I3d98692e69d94756e0034c83a247e05d85177f02