Ensure plugins run before extra.d
Before the code in the plugins was removed from the devstack tree, they
ran before any files in extra.d. When this code is decomposed into a
plugin this order needs to remain the same. Currently this causes
problems with the Ironic devstack plugin and Tempest because the code is
run in this order:

 1. The tempest configuration is run from extra.d, processing
    DEFAULT_INSTANCE_TYPE, and writing the flavor_ref to tempest.conf

 2. The Ironic devstack plugin is run, creating the flavor needed for
    DEFAULT_INSTANCE_TYPE

This leads to build failures as tempest can not find the required
flavor, so it writes which ever flavor it can find at the time into
flavor_ref.

Change-Id: I3d98692e69d94756e0034c83a247e05d85177f02