Ensure plugins run before extra.d
Before the code in the plugins was removed from the devstack tree, they
ran before any files in extra.d, when this code is decomposed into a
plugin this order needs to remain the same. This currently causes
problems with the Ironic devstack plugin and tempest, because the flavor
we need for the DEFAULT_INSTANCE_TYPE is created by the Ironic devstack
plugin, and the tempest extra.d configuration is run before that code
could excute this meant that the flavor_ref in tempest would not match
the DEFAULT_INSTANCE_TYPE, because it didn't exist when that code was
run.

Change-Id: I3d98692e69d94756e0034c83a247e05d85177f02