Switch to od instead of hexdump to allow generate_hex_string to work
For devstack installs that only contain a subset of services,
for example, Heat and Keystone, the hexdump utility is used by
the devstack common function generate_hex_string. The cli is
not part of the general package installs.

This function is used by Heat and Nova to generate values for their conf files.
Using od and tr will avoid the package dependency issues.
Added tests.

Change-Id: I77c8c2019fb8b8174cdfaed3e56ebf728f0732b7
Closes-Bug: #1558672