Update patch set 8
Patch Set 8: Code-Review-1

so i don't see how we're missing hexdump.  this seems to be provided by bsdmainutils which i can see being installed in the image we use for openstack ci at [1]

therefore, we should just add the dependency so that we have hexdump in devstack.  otherwise, we commit stuff that works in the gate (i.e. some separate use of hexdump) and it works, but an external party without it gets a failure.  this might fix up one immediate problem, but leaves us exposed to more.

[1] http://nodepool.openstack.org/dib.ubuntu-trusty.log

Patch-set: 8
Label: Code-Review=-1