Add colorful PS4
PS4 can include functions, so when running in the LOG_COLORS=True mode
provide a grey function line so that it's easier to visually
distinguish the content from the location. Also printf 40 columns
worth of this to make the command lines line up a bit better in the
common case.

Change-Id: I4a10a49db97d413349bcfceeb8c4164936fbcc40