Update patch set 2
Patch Set 2:

Jordan can you provide some examples of projects that might be broken by this so we can work out a solution, the ones that are co-gating with devstack here are unaffected. We really need this for the Ironic devstack plugin, flavor_ref is just an example of one variable that's affected by this, there are others and we do not want to have to duplicate a whole bunch of what extra.d/tempest already does in our plugin.

Patch-set: 2