Update patch set 2
Patch Set 2:

It's going to be an endless battle. If we change this, a couple of weeks later someone is going to propose a similar patch that will undo this one.

I suggest that since Ironic plugin creates the "baremetal" flavor, then Ironic plugin should do the "iniset tempest.conf ..." call.

The other approach would be to make Tempest a devstack plugin, and not an "extras.d" hook as it is right now. I am not sure why it didn't happen yet.

I am sorry I still think this is not a good idea. Let's wait for devstack core to comment.

Patch-set: 2