Update patch set 7
Patch Set 7:

Jordan, it will be a finite number of deployment options. Ideally it should be around 2 or 3. We've proposed removing support for eventlet -- https://review.openstack.org/#/c/297324/ - so that is already one less.

We need devstack to support these deployment options if we actually want to test and support these recommended patterns. We aim to have our CI run tests with multiple deployments. Whether support for running these options exists in the devstack repo or as devstack plugins in the keystone repo is up for debate, I'd prefer we keep things in devstack since that's where it is today.

Patch-set: 7
Label: Verified=0