Allow fetching IPA ramdisk with branch name
This adds branch name to the default URLs for IPA images, ensuring that
devstack runs will use the correct branch of IPA for tests.

It introduces a variable IPA_DOWNLOAD_BRANCH to accomplish this, which
defaults to master. This default will be set to the branch name on
stable branches, e.g. stable/mitaka. It is modified to match our URL
scheme for agent images, e.g. stable-mitaka.

This patch uses stable/liberty as the default, and is an indirect
backport of ironic commit 8e15d6a1388f49026c4d35ba87219a49ccc3c134, as
ironic moved to a devstack plugin in Mitaka.

Change-Id: I045f1a8d674acae9f45bdde8692d4ecc19b328c8
Depends-On: I8c767fa1b56a566c1acabab8552a51ffeb4e1d7d
Closes-Bug: #1561756