Allow fetching IPA ramdisk with branch name
This adds branch name to the default URLs for IPA images, ensuring that
devstack runs will use the correct branch of IPA for tests.

It introduces a variable IPA_DOWNLOAD_BRANCH to accomplish this, which
defaults to master. This default will be set to the branch name on
stable branches, e.g. stable/mitaka. It is modified to match our URL
scheme for agent images, e.g. stable-mitaka.

This patch uses stable/liberty as the default, and is an indirect
backport of ironic commit cc5ac3f98a2e8e518db57bb1eb95db404e11d922, as
ironic moved to a devstack plugin in Mitaka.

Change-Id: I045f1a8d674acae9f45bdde8692d4ecc19b328c8
Depends-On: I8c767fa1b56a566c1acabab8552a51ffeb4e1d7d
Closes-Bug: #1561756