Use Apache/mod_proxy as TLS proxy instead of stud
A template is used to generate configuration for each
service and is appended to tlsproxy.conf in the
appropriate Apache config subdir. This is still
readable and makes cleanup simpler.

A new function was created, add_tls_proxy, because
some services setup multiple listeners and this avoids
Apache restart churn.

No SSLCertificateChainFile/SSLCACertificateFile is set
because the devstack CA is added to the system bundle.

Basic URL rewriting of json content data is done
because some of the services are not proxy-aware.

The SSL/TLS certificate generated by devstack is now
broken out into separate cert/key files for use by
Apache.

Change-Id: Iafb2fa24b7738779c82171dfc57d43877c2e5ed6