Move service availbility iniset to beginning of tempest config
This commit moves the code block which sets the available services in
the tempest config to the beginning of the config function. This is
necessary because some of the tempest utilities which are run depend
on the available services to determine which actions to perform. This
caused an issue with dsvm jobs using accounts.yaml where the networks
weren't being created prior to tempest being run and all tests that
required a network would fail.

Change-Id: I5844aed4e134fbc7210aa0eca83500e260915b7b