Update patch set 3
Patch Set 3:

This is kinda why I don't really think coming up with a really good set of flavors here is a useful exercise. I often have only 10-20G available on my hosts, so I want to start instances with as little disk as possible, which is like 5G for me for a full distro and less for a toy one. I often want to start with lots of memory if I'm trying to perf test something or create obvious scheduler retries, etc.

So, personally, I'd just prefer to create the flavors like we had before so we don't break anyone and expect them to create things that work (like tempest does).

Patch-set: 3