Update patch set 1
Patch Set 1:

Hi Yuriy, 

I have concerns about the name. From 
https://review.openstack.org/#/c/212501/9..12/tempest/api/volume/admin/test_volumes_backup.py
The import test is skipped when the value is True.

    @testtools.skipIf(CONF.volume_feature_enabled.backup_import,
                      "Skip. Bug #1476416 is not fixed on Kilo.")

We set the value True in Kilo. As a result the test is skipped for Kilo. 

Should it be skip_backup_import batter? 

Sorry if I made a mistake here.

Patch-set: 1