Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I'm not really sure what the point is, because they specify "python" as the interpreter and thus this won't get any testing.  we "fix" it and it just goes back to bit-rotting.

there is a real bug in the returning None bit

At least let's not bring in python2.6 workarounds that aren't necessary, we (and OpenStack) don't support that

Patch-set: 4
Label: Code-Review=-1