Update patch set 2
Patch Set 2:

Every 6 months the defaults here change. That part of the code is not stable and we keep coming up with new "ways" to deploy services, so I find that documenting the default behavior is not super useful and will soon be inaccurate.

Anyway I am not opposed to doing it, so +0.

Patch-set: 2