Always create the Fernet repo in prep for the changeover from UUID
As Keystone changes from UUID to Fernet, the default will be changing.
On a given run, we don;t know which wil be the default.  However,
if the default is set to Fernet, and the Key Repository is not
created, Keystone fails to deploy properly. Thuis, always create the Key
repository.

Change-Id: If3cfd53cdf146e18f368b3ee0c45210d4216ad2a