Update patch set 3
Patch Set 3: Code-Review-1

We don't mention Mint, or Debian, or SuSE which are all other things people use, and Fedora 22 likely doesn't work.  i.e. this list is not helpful and i don't see this makes it better.

Trying to be constructive: what we could do here is split this into to categories.  We can say that the officially supported environments are those that run devstack jobs in OpenStack CI which is currently ubuntu-trusty, centos7 & fedora23.  Then have a second category "reported to work on" which would include RHEL7, OracleLinux7, SuSe, Mint & Debian (at least, maybe more)

Patch-set: 3
Label: Code-Review=-1