neutron: Set privsep.helper_command
When neutron starts using privsep, it will need to know how to invoke
its privileged half.  Amazingly the name of the rootwrap executable
isn't anywhere else in the config, so the privsep default uses just
"sudo" (no rootwrap).

We need to either:

1. set the privsep command line to use neutron-rootwrap in
   neutron.conf (and similar in other configs), or

2. add the privsep-helper line to sudoers and bypass rootwrap entirely.

This change implements (1) for devstack/neutron.

An additional complication with neutron is that we need to support the
wide variety of "--config-file" arguments employed by neutron's devstack
logic.

Change-Id: Ia9675dff9232e0e987a836ecaf9e842eb5c3cb18