Update patch set 1
Patch Set 1:

This is just allowing the overriding of a variable at a different point in the devstack run.  Typically you use tools/xen/install_os_domu as the first entry point, which already allows configuring of this setting (through XEN_INT_BRIDGE_OR_NET_NAME).
In this case, the change is needed to allow moving of the VM to another host (with a different XEN_INT_BRIDGE) for easier install.

BTW - configuration.rst says that XenServer specific docs would be in tools/xen/README.md.

Patch-set: 1