Update patch set 2
Patch Set 2:

@Dean: Screen presents tremendous value in debugging and restarting services.  While stackenv gives the list of services, it does not provide the restart information as to what config files were specified to start the services. 

As to unstack/stack being the only way to restart services, this defeats the purpose where snapshots are used instead of repeatedly running unstack/stack. For instance once the system is up and running, we take a snapshot and save it and we can revert to it anytime - but, with this patch we loose that capability - this forces us to run unstack/stack in order to bring the systems back on-line. 
If you look at the statistics from all third party CIs 50+% failures are attributed to the failures of stack.sh. Using snapshot instead of running unstack/stack every time makes the systems reliable as well as saves tons of time, as unstack/stack is very expensive operation. 

Moreover, this requires many vendors to go back and change their scripts - a tremendous resource consumption for no obvious gain. 

I hope you will be sensitive to such constraints.

Patch-set: 2